requirements.txt
Flask
Flask-SocketIO
gunicorn
requests
apscheduler
finnhub-python
yfinance
